requirements.txt
keyboard==0.13.5
openai==1.1.2
PyAutoGUI==0.9.54
pygame==2.5.2
python-dotenv==1.0.0
Requests==2.31.0
SpeechRecognition==3.10.0
